perf(web): keep react-virtualized out of the entry chunk
react-virtualized is ~638 KB of source and was landing in the entry chunk for
every visitor through two unrelated edges, neither of which needs it at first
paint.

1. harmony's FilterButton imported it statically, but only renders
   VirtualizedOptionsList behind the opt-in `virtualized` prop -- 3 call sites
   across the whole web app. Split into its own module and loaded on demand
   behind a Suspense boundary. The menu only renders while open, so the import
   starts when a virtualized filter menu is actually opened.

   Note the `as typeof import(...)` on the lazy const: React.lazy erases the
   component's `<Value extends string>` generic, so it would otherwise resolve
   to the `string` instantiation and lose type safety at call sites.

2. TransactionDetailsContent imported `isChangePositive` -- a five-line
   predicate over two enum values -- from AudioTransactionsTable. That module
   imports components/table, which imports react-virtualized. Because the
   TransactionDetails modal is registered eagerly in registerNiceModals, the
   whole virtualized table came along for a function that checks whether a
   transaction increases a balance. Moved the predicate to its own module and
   re-exported it from the table for backwards compatibility.
